pyproject.toml
[tool.poetry]
name = "supplierserver"
version = "0.1.0"
description = ""
authors = ["Samiiz <skboss159@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pyjwt = "^2.10.1"
black = "^24.10.0"
isort = "^5.13.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
